Fix static issues in AudioD
:Release Notes:
Fix static issues related to resource leaks and argument type
mismatch in Audiod

:Detailed Notes:
Addressed below static issues
CID-8481185 Resource leak
CID-8481184 Resource leak
CID-8480504 Printf arg type mismatch

:Testing Performed:
Tested on RPI4 rebased on OSE build #251
Test case: QWA-4099 QWA-3810 QWA-5429

:QA Notes:
NA

:Issues Addressed:
[CID-8481185] Resource leak
[CID-8481184] Resource leak
[CID-8480504] Printf arg type mismatch

+++ src/modules/audioPolicyManager/audioPolicyManager.cpp
@@ -332,7 +332,7 @@ void AudioPolicyManager::applyVolumePolicy(EVirtualAudioSink audioSink, const st
 {
     PM_LOG_INFO(MSGID_POLICY_MANAGER, INIT_KVCOUNT,\
         "AudioPolicyManager:applyVolumePolicy sink:%d streamType:%s priority:%d",\
-        audioSink, streamType.c_str(), priority);
+        (int)audioSink, streamType.c_str(), priority);
     if (mObjAudioMixer)
     {
         int policyPriority = MAX_PRIORITY;
